fix: fix db connections for user-api service
