+++ services/user-api/src/services/stockService.ts
@@ -1,18 +1,32 @@
-import type { Stock } from "shared-models/redisSchema";
+import { Repository } from "redis-om";
+import { RedisInstance } from "shared-models/RedisInstance";
 import {
-  ownedStockRepository,
-  redisConnection,
-  setupRedisConnection,
-  stockRepository,
-  stockTransactionRepository,
-} from "./userApiService";
+  ownedStockSchema,
+  type Stock,
+  type StockOwned,
+  stockSchema,
+  type StockTransaction,
+  StockTransactionSchema,
+} from "shared-models/redisSchema";
+
+// Creating connection here due to the implementation of RedisInstance.ts
+// This is probably NOT good - causes multiple connection creation?
+let redisConnection: RedisInstance = new RedisInstance();
+try {
+  redisConnection.connect();
+} catch (error) {
+  throw new Error("Error starting database server");
+}
+const stockRepository: Repository<Stock> = await redisConnection.createRepository(stockSchema);
+const ownedStockRepository: Repository<StockOwned> = await redisConnection.createRepository(
+  ownedStockSchema
+);
+const stockTransactionRepository: Repository<StockTransaction> =
+  await redisConnection.createRepository(StockTransactionSchema);
 
 const stockService = {
   async createStock(stock_name: string) {
-    if (!redisConnection) {
-      await setupRedisConnection();
-    }
-    const existingStock = await stockRepository!
+    const existingStock = await stockRepository
       .search()
       .where("stock_name")
       .equals(stock_name)
@@ -28,11 +42,8 @@ const stockService = {
     return saved_stock.stock_id;
   },
   async getUserStockPortfolio(userName: string) {
-    if (!redisConnection) {
-      await setupRedisConnection();
-    }
     try {
-      const userStocks = await ownedStockRepository!
+      const userStocks = await ownedStockRepository
         .search()
         .where("user_name")
         .equals(userName)
@@ -43,11 +54,8 @@ const stockService = {
     }
   },
   async getUserStockTransactions(userName: string) {
-    if (!redisConnection) {
-      await setupRedisConnection();
-    }
     try {
-      const userStockTransactions = await stockTransactionRepository!
+      const userStockTransactions = await stockTransactionRepository
         .search()
         .where("user_name")
         .equals(userName)
